updated geo json colors
Update on what I worked on: - fixed div toggle function which made the text disappear when the legend appeared. - im so sorry @Harvey Ucl but one of the wards is missing in the GeoJson (st pancras)-updated buttons and legend css to make the prettier (font and roundness and position)-changed the color geojson to scheme that fits better- I dont know what info the json is displaying but it should be mentioned in the legend up top with the number (im not sure what these numbers are and in what unit). I am considering moving the text divs to the left of the map though that would be significant work at this stage if we deem it necessary. I also reduced the height of the map slightly.